% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init_algo.R
\name{Cluster.by.PCA}
\alias{Cluster.by.PCA}
\title{Clustering by PCA, version 1}
\usage{
Cluster.by.PCA(dat, m.star = 4, K, cutoff.frac = 0.1, verbose = FALSE)
}
\arguments{
\item{dat}{n x p matrix}

\item{m.star}{the number of principal components to consider, should be no bigger than 5}

\item{K}{the number of cluster points to return}

\item{cutoff.frac}{the level on the kernel density below which peaks are not considered}

\item{verbose}{boolean}
}
\value{
candidate matrix
}
\description{
similar to Cluster.by.PCA(..)
}

