% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init_algo.R
\name{data2cluster}
\alias{data2cluster}
\title{Assigns a cluster number (index) to each data point}
\usage{
data2cluster(Cpts, dat)
}
\arguments{
\item{Cpts}{n.ctr x p matrix of cluster points returned by Clusters.by.PCA, 
where each row is a point; columns are dimension of the data}

\item{dat}{n x p matrix of the data, rows are observation and columns dimension}
}
\value{
vector of the same length as the number of observations, which tells 
         you which cluster center the data point is closest to.
}
\description{
Assigns a cluster number(index) to a data point, where the center
    of the clusters is given by Cpts.
}

