% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init_algo.R
\name{init.GM.by.kmeans}
\alias{init.GM.by.kmeans}
\title{Initialization by k-means method}
\usage{
init.GM.by.kmeans(obs, K, M, Pi = NULL, A = NULL, dir.a = 1,
  var.reduct = 1, verbose = TRUE)
}
\arguments{
\item{obs}{data used in hextfit(..), must have components : x$x and x$TT}

\item{K}{number of hidden state}

\item{M}{number of gaussian mixtures}

\item{dir.a}{dir.a}

\item{var.reduct}{variance reduction factor}
}
\value{
list of Pi, initial probability of the hidden states, A, the transition matrix, and b, the emission matrix
}
\description{
This methods runs k-means twice, first to segment the entire data set into K groups,
      then segment each group into M subgroups corresponding to the mixture components.
}

