% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init_algo.R
\name{init.GM.by.LL}
\alias{init.GM.by.LL}
\title{Best Loglikelihood algorithm}
\usage{
init.GM.by.LL(x, K, M, D, mtype, Nset = 10, dir.a = 1, var.reduct = 1,
  verbose = TRUE)
}
\arguments{
\item{x}{data used in hextfit(..), must have components : x$x and x$TT}

\item{K}{number of hidden state}

\item{M}{number of gaussian mixtures}

\item{D}{vector of the size of the alphabets in each column of the multivariate discrete obs matrix}

\item{mtype}{vector to indicate the type of variables in obs}

\item{Nset}{number of random sets from which to choose the best one}

\item{dir.a}{dir.a}

\item{var.reduct}{variance reduction factor}
}
\value{
list of Pi, initial probability of the hidden states, A, the transition matrix, and b, the emission matrix
}
\description{
Choose initial values by picking the best set from Nset sets of random 
        values. This function calls init.GM.random to generate Nset random sets.
}

