% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init_algo.R
\name{init.GM.by.PCA}
\alias{init.GM.by.PCA}
\title{Initialization by PCA}
\usage{
init.GM.by.PCA(x, K, M, Pi = NULL, A = NULL, spherical = TRUE,
  var.reduct = 1, fudge = 0.1, dir.a = 1, verbose = FALSE)
}
\arguments{
\item{x}{data used in hextfit(..), must have components : x$x and x$TT}

\item{K}{number of hidden state}

\item{M}{number of gaussian mixtures}

\item{var.reduct}{variance reduction factor}

\item{dir.a}{dir.a}

\item{D}{vector of the size of the alphabets in each column of the multivariate discrete obs matrix}

\item{mtype}{vector to indicate the type of variables in obs}

\item{Nset}{number of random sets from which to choose the best one}
}
\value{
list of Pi, initial probability of the hidden states, A, the transition matrix, and b, the emission matrix
}
\description{
This algorithm uses principal component analysis to discern local maxima 
in the data. It divides the data into K equal sized segments, then send
each segment into Cluster.by.PCA to find M clustering points in the data
space, which becomes the mu's of the gaussian mixtures in M5. The 
covariance matrices of the components are calculated from the PCA-segmented
data whenever possible, and are made spherical if the segmented data is
insufficient to come up with a positive definite covariance matrix. If the 
number of points returned by Cluster.by.PCA is less than M, then the deficit
is supplied with random vectors centered around the PCA-clustered points.
}

