% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init_algo.R
\name{init.GM.max.sep}
\alias{init.GM.max.sep}
\title{Initialization by Max Sep Algorithm}
\usage{
init.GM.max.sep(obs, K, M, Pi = NULL, A = NULL, dir.a = 1,
  var.reduct = 1, verbose = TRUE)
}
\arguments{
\item{obs}{data used in hextfit(..), must have components : x$x and x$TT}

\item{K}{number of hidden state}

\item{M}{number of gaussian mixtures}

\item{dir.a}{dir.a}

\item{var.reduct}{variance reduction factor}
}
\value{
list of Pi, initial probability of the hidden states, A, the transition matrix, and b, the emission matrix
}
\description{
This is the maximal separation algorithm described in my paper (cite)
      which finds the points farthest away from each other. It calls max.separate.set
}
\details{
Note that M5 has the same structure as in gen.init.
}

