% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init_algo.R
\name{init.GM.random}
\alias{init.GM.random}
\title{Random initialization for Pi, A, and b}
\usage{
init.GM.random(obs, K, M, var.reduct = 1, dir.a = 1, verbose = FALSE)
}
\arguments{
\item{K}{number of hidden state}

\item{M}{number of gaussian mixtures}

\item{x}{data used in hextfit(..), must have components : x$x and x$TT}

\item{D}{vector of the size of the alphabets in each column of the multivariate discrete obs matrix}

\item{mtype}{vector to indicate the type of variables in obs}
}
\value{
list of Pi, initial probability of the hidden states, A, the transition matrix, and b, the emission matrix
}
\description{
The means of the gaussian mixture are generated by uniform sampling 
              from the observed range in each dimension. The variances for each
              spherical covariance matrix is picked uniformly from half the 
              min variance to the maximum variance.
}

