% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init_algo.R
\name{init.GM.Seg.KMeans}
\alias{init.GM.Seg.KMeans}
\title{Segmental k-means algorithm}
\usage{
init.GM.Seg.KMeans(x, K, M, D, mtype, Pi = NULL, A = NULL, tol = 0.01,
  maxit = 20, dir.a = 3, var.reduct = 1, verbose = FALSE)
}
\arguments{
\item{x}{data used in hextfit(..), must have components : x$x and x$TT}

\item{K}{number of hidden state}

\item{M}{number of gaussian mixtures}

\item{D}{vector of the size of the alphabets in each column of the multivariate discrete obs matrix}

\item{mtype}{vector to indicate the type of variables in obs}
}
\value{
list of Pi, initial probability of the hidden states, A, the transition matrix, and b, the emission matrix
}
\description{
see my paper for detailed description
}
\details{
Reference : Juang & Rabiner, Mixture Autoregressive HMM for Speech Signals, IEEE
          Transactions on Acoustics, Speech, and Signal Processing, Vol. ASSP-33, No 6, Dec 1985.
}

