% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init_algo.R
\name{init.GM.unif.seg}
\alias{init.GM.unif.seg}
\title{Uniform segmentation method}
\usage{
init.GM.unif.seg(x, K, M, Pi = NULL, A = NULL, var.reduct = 1,
  dir.a = 1, verbose = TRUE)
}
\arguments{
\item{x}{data used in hextfit(..), must have components : x$x and x$TT}

\item{K}{number of hidden state}

\item{M}{number of gaussian mixtures}

\item{var.reduct}{variance reduction factor}

\item{dir.a}{dir.a}

\item{D}{vector of the size of the alphabets in each column of the multivariate discrete obs matrix}

\item{mtype}{vector to indicate the type of variables in obs}

\item{Nset}{number of random sets from which to choose the best one}
}
\value{
list of Pi, initial probability of the hidden states, A, the transition matrix, and b, the emission matrix
}
\description{
This uniform segmentation method is described in the HTK Toolkit (Section 8.2, p.133)
and makes sense only for left-right models. The codes below check whether a 
covariance matrix is positive definite, if not it computes the nearest p.d. matrix
using the algorithm in NJ Higham (1988, Linear Algebra Appl. 103:103-118).
}

