% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init_algo.R
\name{init.GM.unifcommon}
\alias{init.GM.unifcommon}
\title{Uniform common start method}
\usage{
init.GM.unifcommon(x, K, M, Pi = NULL, A = NULL, var.reduct = 1,
  dir.a = 1, spherical = TRUE, verbose = TRUE)
}
\arguments{
\item{x}{data used in hextfit(..), must have components : x$x and x$TT}

\item{K}{number of hidden state}

\item{M}{number of gaussian mixtures}
}
\value{
list of Pi, initial probability of the hidden states, A, the transition matrix, and b, the emission matrix
}
\description{
see my paper for detailed description
}
\details{
Reference : HTK Toolkit, Ch. 8, HINT module
}

