% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init_algo.R
\name{Kmeans.Vector.Quantization}
\alias{Kmeans.Vector.Quantization}
\title{Vector quantization by k-means}
\usage{
Kmeans.Vector.Quantization(obs, K)
}
\arguments{
\item{obs}{observation matrix}

\item{K}{number of codewords}
}
\value{
list of codebk, the codebook, mu.m, the cluster center, and sigma.l, the variance of the clusters.
}
\description{
Specifying a K, apply k-means to data. The center of the clusters is mu, while the cluster 
       index is returned as the codebook.
}

