% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init_experiments.R
\name{LL.plot}
\alias{LL.plot}
\title{Plot the loglikelihood trace of the fitted HMM}
\usage{
LL.plot(mod, n.trunc = 5)
}
\arguments{
\item{mod}{HMM model}

\item{n.trunc}{n.trunc}
}
\value{
None
}
\description{
plot the LL
}

