% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init_experiments.R
\name{loop.over.hext.plot}
\alias{loop.over.hext.plot}
\title{Plotting function for 2-dim models}
\usage{
loop.over.hext.plot(init.method, x, mtype, K, K.truth, M, D, wdf = 1,
  tol = 0.1, maxit = 1000, var.reduct = 0.5, fudge = 0.1,
  spherical = TRUE, verbose = TRUE, fname = NULL)
}
\arguments{
\item{init.method}{initialization method}

\item{x}{data matrix for Hext models, where each row is an observation at time t, columns are features}

\item{mtype}{vector to indicate the type of variables in obs}

\item{K}{vector of length Nruns, each component is the number of states for the ith run}

\item{M}{vector of length Nruns, each component is the number of mixtures for the ith run}

\item{D}{vector of the size of the alphabets in each column of the multivariate discrete obs matrix}

\item{wdf}{degree of freedom of the Wishart}

\item{tol}{tolerance}

\item{maxit}{maximum number of iterations for each run}

\item{var.reduct}{variance reduction factor}

\item{fudge}{fudge factor}

\item{spherical}{boolean}

\item{verbose}{boolean}

\item{mod}{HMM model}

\item{TT}{vector of the lengths of observation sequences}
}
\value{
None
}
\description{
Plot function to write the plot to a PDF file. See loop.over.hext(..) for details.
}

