% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init_experiments.R
\name{loop.over.hext}
\alias{loop.over.hext}
\title{Loop over the single.run function}
\usage{
loop.over.hext(init.method, x, mtype, K, M, D, maxit = 1000, tol = 1e-04,
  wdf = 1, var.reduct = 0.5, fudge = 0.1, verbose = TRUE,
  spherical = TRUE)
}
\arguments{
\item{init.method}{initialization algorithm to use}

\item{x}{data matrix for Hext models, where each row is an observation at time t, columns are features}

\item{mtype}{vector to indicate the type of variables in obs}

\item{K}{vector of length Nruns, each component is the number of states for the ith run}

\item{M}{vector of length Nruns, each component is the number of mixtures for the ith run}

\item{D}{vector of the size of the alphabets in each column of the multivariate discrete obs matrix}

\item{maxit}{maximum number of iterations for each run}

\item{tol}{tolerance}

\item{wdf}{degree of freedom for Wishart}

\item{var.reduct}{variance reduction factor}

\item{fudge}{fudge factor}

\item{verbose}{print out status of the run, boolean}

\item{spherical}{spherical}
}
\value{
list of Lik, K, M, sigmal.ll, mu.ll
}
\description{
See single.run(...) for details
}

