% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init_algo.R
\name{make.noisy.copies}
\alias{make.noisy.copies}
\title{Make copies of x by adding noise}
\usage{
make.noisy.copies(x, n, var.s)
}
\arguments{
\item{x}{either a vector of length p or a matrix with p columns}

\item{n}{number of copies to make}

\item{var.s}{scalar variance for spherical multivariate noise}
}
\value{
n x p matrix where each column is a noisy copy of x
}
\description{
See code
}

