% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init_experiments.R
\name{match.confusion}
\alias{match.confusion}
\title{Measure the decoding performance of a hidden Markov model}
\usage{
match.confusion(ss, ss.fit, digit = 3)
}
\arguments{
\item{ss}{true sequence of hidden states}

\item{ss.fit}{predicted state sequence from a decoding algorithm, e.g. Viterbi}

\item{digit}{number of digit, default 3}
}
\value{
a matrix of size K x K.out, where K is the number of true states in the obs sequence, K.out is the number of states
        decoded by the decoding algorithm. It could be interpreted as follow. If the decoding is perfect,
         then every row and every column would have *exactly* one non-zero element. This means every actual
         state is mapped to *exactly* one decoded state. However, if a row has multiple entries, 
         that means the HMM thinks some of state i look different and assign them to more than
         one state. If a column has multiple entries, that means the variability of 
         certain state is so great that some of its observations are classified into separate
         decoded states.
}
\description{
If the decoding is perfect, then the decoded state sequence should match with the true state
         sequence exactly. If some of the observations should be in state i, but is mapped to 
         state j, then that is a decoding error.
}

