% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init_experiments.R
\name{mu.trace}
\alias{mu.trace}
\title{Plot the trajectories of the gaussian means over each EM iterations until convergence}
\usage{
mu.trace(dat, n.it, K, M, LLtrace, mu)
}
\arguments{
\item{dat}{data set}

\item{n.it}{number of iterations, to be use to loop over mu}

\item{K}{number of hidden state}

\item{M}{number of gaussian mixtures}

\item{LLtrace}{LLtrace}

\item{mu}{list with the same structure as model$parms.emission$M5$mu.ll, see details}
}
\value{
None
}
\description{
paramter mu is a list where each element holds the paramter of one iteration, and each element 
             has the gaussian means of each mixture component in each state.
}

