% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init_experiments.R
\name{mu.trajectory}
\alias{mu.trajectory}
\title{Records the position of the means of each Gaussian mixture state for clustering display}
\usage{
mu.trajectory(Nruns, init.method, x, mtype, K, M, D, maxit = 1000,
  tol = 1e-04, wdf = 1, var.reduct = 0.5, fudge = 0.1, verbose = TRUE,
  spherical = TRUE)
}
\arguments{
\item{Nruns}{number of runs}

\item{init.method}{initialization method}

\item{x}{data structure for Hext models}

\item{mtype}{vector to indicate the type of variables in obs}

\item{K}{number of hidden states}

\item{M}{number of mixtures for each state}

\item{D}{vector of the size of the alphabets in each column of the multivariate discrete obs matrix}

\item{maxit}{maximum number of iterations, default = 1000}

\item{tol}{tolerance, default = 1e-14}

\item{wdf}{degree of freedom for Wishart}

\item{var.reduct}{var.reduct}

\item{fudge}{fudge factor}

\item{verbose}{boolean}

\item{spherical}{boolean}

\item{TT}{length of observation sequence, if vector, it has the length of each sequence in x}
}
\value{
list of three elements : mu.init, the initial position of the gaussian means, mu.final, the final position,
           ans, the result of one call to single.run(..)
}
\description{
see example
}

