% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init_algo.R
\name{PCA.algo.test}
\alias{PCA.algo.test}
\title{Test the PCA algorithm}
\usage{
PCA.algo.test(obs, K, M)
}
\arguments{
\item{obs}{observation matrix}

\item{K}{K}

\item{M}{M}
}
\value{
list of Centers, Memb, where Centers is a list of K lists, each being the center returned by Cluster.by.PCA
}
\description{
Call Cluster.by.PCA multiple (K) times
}

