% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init_algo.R
\name{pk.find}
\alias{pk.find}
\title{Locates the maximum value in a vector x and returns the location.}
\usage{
pk.find(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
vector of peak locations
}
\description{
Low level function being called in a rollapply loop by pk.scoring
}

