% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init_algo.R
\name{pk.scoring}
\alias{pk.scoring}
\title{Internal function to show the score from the first pass}
\usage{
pk.scoring(z, win.width = 21)
}
\arguments{
\item{z}{a zoo object}

\item{win.width}{window size}
}
\value{
vector of votes, same length as z
}
\description{
Note requires rollapply from zoo package
}

