% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init_experiments.R
\name{single.run}
\alias{single.run}
\title{One pass through the HMM fit-and-decode cycle}
\usage{
single.run(x, mtype, K, M, D, init.method = "random", wdf = 1, tol = 0.1,
  maxit = 1000, var.reduct = 1, fudge = 0.1, spherical = TRUE,
  verbose = TRUE)
}
\arguments{
\item{x}{data structure for Hext models}

\item{mtype}{vector to indicate the type of variables in obs}

\item{K}{number of states for the run}

\item{M}{number of mixtures for the run}

\item{D}{vector of the size of the alphabets in each column of the multivariate discrete obs matrix}

\item{init.method}{initialization method}

\item{wdf}{degree of freedom for Wishart}

\item{tol}{tolerance}

\item{maxit}{maximum number of iterations}

\item{var.reduct}{variance reduction factor}

\item{fudge}{fudge factor}

\item{spherical}{spherical}

\item{verbose}{boolean}

\item{mod}{HMM model}

\item{TT}{vector of the lengths of observation sequences}
}
\value{
list of the following items :
         sseq : the decoded state sequences,
         LL : loglikelihood of the HMM model,
         AIC, AICc, BIC : the AIC, AICC, and BIC of the model,
         init : the initial value of the paramters generated by the initialization method,
         res : the fitted model,
         last.it : the state of last iteration.
}
\description{
This function takes the observation sequences, initialize the HMM parameters, call hextfit to 
             fit a model, then decode the hidden state sequence with the Viterbi algorithm.
}

