% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init_experiments.R
\name{Vit.SS.plot}
\alias{Vit.SS.plot}
\title{Plot the Viterbi states sequence by sequence}
\usage{
Vit.SS.plot(SS, TT)
}
\arguments{
\item{SS}{state sequences}

\item{TT}{vector of length of each sequence}
}
\value{
None
}
\description{
plot the decoded state sequence
}

